Removed JokeValue class, updated JokeService and JokeResponse
Removed the JokeValue class as it is no longer necessary. The JokeService and JokeResponse classes have been updated to handle the responses directly in a more efficient way. These changes were necessary as the original 'icndb' joke website is no longer available. Now, we use the 'api.chucknorris.io' service for retrieving random developer jokes. This change also affected the JokeServiceTest and application properties, adjusting to the new service. Also updated the URL in the JokeServiceTest file to reflect this change.