"Updated test setup and removed unsupported encoding exception"
Updated the GeocoderServiceTest setup to include an @BeforeEach method with a check for the availability of the Google Maps API and removed an assertion for 'address' on 'getLatLngWithoutStreet' test. Changes to the GeocoderService included the switch from an unsupported encoding exception to the StandardCharset's UTF-8, for the sake of thread-safety and efficiency. Furthermore, a private key was updated.