"Update codebase for Java 17 and Spring Boot 3.2"
This commit is a significant refactoring of the codebase to make use of new Java 17 features and to improve readability and efficiency. The source compatibility is updated to Java 17. With this, the Greetings class has been rewritten as a Record, reducing boilerplate code.

Request parameters in HelloController & HelloRestController are no longer mandatory implying, by default, the name parameter's value is "World".

The 'AppConfig' class no longer creates 'Greeting' beans but instead creates a number of 'NumberFormat' beans. As a consequence of this, affected unit tests are updated to reflect the AppConfig changes, and a test for absent Greeting bean is introduced.

The Gradle wrapper is upgraded to version 8.4, and the Spring Boot version in the build.gradle file was updated to '3.2.0' from '2.7.12', and its corresponding dependencies have been updated. Reorganized the project's unit tests and changed some assertions to match the new functionality. The 'HelloControllerIntegrationTest' class was renamed to 'HelloControllerMockMVCTest'.

The changes aim to keep the project up-to-date with the latest Java & Spring Features and best practices. Also, implementation of the NumberFormat beans are more useful and illustrative than the previous Greeting ones.