Upgraded Spring Boot and other dependencies, refactored code and added new tests.
This commit includes multiple changes primarily focused on upgrading the versions of Spring Boot and other dependencies in the project. With this, we are ensuring compatibility with the latest frameworks and making use of their new features and improvements.
The build.gradle file was changed to reflect these upgraded versions. The java compatibility was also upgraded to version 17.

The code has been refactored to reflect these updates, for instance where we switched from javax.persistence to jakarta.persistence in JpaOfficerDAO.java.

New tests were added in JdbcClientOfficerDAOTest.java, and a new class JdbcClientOfficerDAO.java was also added. These changes are made to increase the coverage of the code and find potential bugs or issues.

The distribution url in gradle-wrapper.properties was also updated to a new one as part of the Gradle upgrade.

Rogue imports were cleaned up in JdbcOfficerDAO.java, likely left from the previous refactorings.

The equals method in Officer.java was simplified using Java's Objects class, reducing potential for error in object comparison.

All of these changes aim to improve the quality and maintainability of the codebase by staying up-to-date with the latest frameworks, improving test coverage and removing unnecessary code.