Refactor JdbcClientOfficerDAO.java to enhance efficiency
The refactoring of JdbcClientOfficerDAO.java was carried out to improve the efficiency of querying the 'officers' table. This involved removing the explicit definition of the RowMapper officerMapper and reducing its utilization to enhance simplicity. Hence, we now utilize 'Officer.class' in our queries instead of employing the RowMapper to map our ResultSet to the Officer object. This change makes the code cleaner and easier to read. The DAO's constructor was also modified to be directly supplied with an instance of JdbcClient, allowing for better testing and control over its usage elsewhere in the application.